+++ src/environments/environment.prod.ts
@@ -1,3 +1,4 @@
 export const environment = {
-  production: true
+  production: true,
+  GOOGLE_API_KEY: 'AIzaSyB_TNfE3rG_JOV-3jrDNT04lml2sxPzNUc'
 };