+++ CHANGELOG.md
@@ -2,6 +2,57 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="0.5.1"></a>
+## [0.5.1](https://gitlab.com/Tott0/Caxtor/compare/v0.5.0...v0.5.1) (2018-05-29)
+
+
+### Bug Fixes
+
+* prod build ([33fde20](https://gitlab.com/Tott0/Caxtor/commit/33fde20))
+
+
+
+<a name="0.5.0"></a>
+# [0.5.0](https://gitlab.com/Tott0/Caxtor/compare/v0.4.0...v0.5.0) (2018-05-29)
+
+
+### Bug Fixes
+
+* stopped using mocks ([55ed6ec](https://gitlab.com/Tott0/Caxtor/commit/55ed6ec))
+
+
+### Features
+
+* readme ([5d6a87e](https://gitlab.com/Tott0/Caxtor/commit/5d6a87e))
+
+
+
+<a name="0.4.0"></a>
+# [0.4.0](https://gitlab.com/Tott0/Caxtor/compare/v0.3.0...v0.4.0) (2018-05-04)
+
+
+### Features
+
+* **client:** warehouse add top filters ([bd37a9f](https://gitlab.com/Tott0/Caxtor/commit/bd37a9f))
+
+
+
+<a name="0.3.0"></a>
+# [0.3.0](https://gitlab.com/Tott0/Caxtor/compare/v0.1.0...v0.3.0) (2018-05-01)
+
+
+### Bug Fixes
+
+* **auth:** fixes signup and login issues ([d2c24d8](https://gitlab.com/Tott0/Caxtor/commit/d2c24d8))
+
+
+### Features
+
+* remade navbar, warehouse creation working, warehouse dashboard ([f93a673](https://gitlab.com/Tott0/Caxtor/commit/f93a673))
+* **owner:** warehouse creation ([3e6d87d](https://gitlab.com/Tott0/Caxtor/commit/3e6d87d))
+
+
+
 <a name="0.2.1"></a>
 ## [0.2.1](https://gitlab.com/Tott0/Caxtor/compare/v0.2.0...v0.2.1) (2018-04-30)